feat(layout-ui)  : create brand form the sidebar
