+++ src/shared/components/Sidebar/CreateNewItemPanel.js
@@ -1,14 +1,17 @@
 import { useMutation } from '@apollo/react-hooks'
+import { Tunnel, Tunnels, useTunnel } from '@dailykit/ui'
 import React from 'react'
 import { toast } from 'react-toastify'
+import CreateBrandTunnel from '../../../apps/brands/views/Listings/brands/CreateBrandTunnel'
 import { GENERAL_ERROR_MESSAGE } from '../../../apps/inventory/constants/errorMessages'
 import { CREATE_SUPPLIER } from '../../../apps/inventory/graphql'
 import { RectangularIcon } from '../../assets/icons'
-import { useTabs } from '../../providers'
+import { useTabs, TooltipProvider } from '../../providers'
 import { logger, randomSuffix } from '../../utils'
 import Styles from './styled'
 
 const CreateNewItemPanel = () => {
+   const [tunnels, openTunnel, closeTunnel] = useTunnel(1)
    const { addTab } = useTabs()
    const [createSupplier] = useMutation(CREATE_SUPPLIER, {
       onCompleted: input => {
@@ -35,13 +38,20 @@ const CreateNewItemPanel = () => {
 
    return (
       <Styles.CreateNewItems>
-         <CreateNewBtn title="Brands" />
+         <CreateNewBtn title="Brands" onClick={() => openTunnel(1)} />
          <CreateNewBtn title="Collection" />
          <CreateNewBtn title="Products" />
          <CreateNewBtn title="Recipe" />
          <CreateNewBtn title="Ingredient" />
          <CreateNewBtn title="Supplier Item" />
          <CreateNewBtn title="Supplier" onClick={createSupplierHandler} />
+         <Tunnels tunnels={tunnels}>
+            <Tunnel layer={1} size="md">
+               <TooltipProvider app="Brand App">
+                  <CreateBrandTunnel closeTunnel={closeTunnel} />
+               </TooltipProvider>
+            </Tunnel>
+         </Tunnels>
       </Styles.CreateNewItems>
    )
 }
